vim: Migrate from Pathogen to Vundle, add plugins.
This commit does a couple of things, we migrate from Pathogen to
Vundle, since the latter seems to be a bit easier to work with. We
also drop out vim-powerline for vim-airline, since the former is no
longer supported, and its supposed replacement is a bloated mess of
feature creep. Airline does one thing well, and that's provide a set of
status line hooks in a few hundred lines of vimscript.

We also bring in NERDTree, vim-go, and vim-gitgutter.